[SOIN] Déplace les routes de la visite guidée dans un routeur dédié
