+++ src/routes/privees/routesApiPrivee.js
@@ -25,6 +25,7 @@ const {
 const {
   verifieCoherenceDesDroits,
 } = require('../../modeles/autorisations/gestionDroits');
+const routesApiVisiteGuidee = require('./routesApiVisiteGuidee');
 
 const routesApiPrivee = ({
   middleware,
@@ -137,6 +138,16 @@ const routesApiPrivee = ({
       adaptateurPdf,
       adaptateurZip,
       adaptateurJournalMSS,
+    })
+  );
+
+  routes.use(
+    '/visiteGuidee',
+    middleware.verificationAcceptationCGU,
+    routesApiVisiteGuidee({
+      middleware,
+      depotDonnees,
+      referentiel,
     })
   );
 
@@ -397,78 +408,6 @@ const routesApiPrivee = ({
           initiales: c.initiales(),
         })),
       });
-    }
-  );
-
-  routes.post(
-    '/visiteGuidee/:idEtape/termine',
-    middleware.verificationAcceptationCGU,
-    middleware.aseptise('idEtape'),
-    async (requete, reponse) => {
-      const { idUtilisateurCourant } = requete;
-      const { idEtape } = requete.params;
-
-      if (!referentiel.etapeVisiteGuideeExiste(idEtape)) {
-        reponse.status(400).send("Identifiant d'étape inconnu");
-        return;
-      }
-
-      const parcoursUtilisateur =
-        await depotDonnees.lisParcoursUtilisateur(idUtilisateurCourant);
-
-      parcoursUtilisateur.etatVisiteGuidee.termineEtape(idEtape);
-      await depotDonnees.sauvegardeParcoursUtilisateur(parcoursUtilisateur);
-
-      const idEtapeSuivante = referentiel.etapeSuivanteVisiteGuidee(idEtape);
-      const urlEtapeSuivante =
-        referentiel.etapeVisiteGuidee(idEtapeSuivante)?.urlEtape ?? null;
-
-      reponse.send({ urlEtapeSuivante });
-    }
-  );
-
-  routes.post(
-    '/visiteGuidee/termine',
-    middleware.verificationAcceptationCGU,
-    async (requete, reponse) => {
-      const { idUtilisateurCourant } = requete;
-
-      const parcoursUtilisateur =
-        await depotDonnees.lisParcoursUtilisateur(idUtilisateurCourant);
-      parcoursUtilisateur.etatVisiteGuidee.finalise();
-      await depotDonnees.sauvegardeParcoursUtilisateur(parcoursUtilisateur);
-
-      reponse.sendStatus(200);
-    }
-  );
-
-  routes.post(
-    '/visiteGuidee/metEnPause',
-    middleware.verificationAcceptationCGU,
-    async (requete, reponse) => {
-      const { idUtilisateurCourant } = requete;
-
-      const parcoursUtilisateur =
-        await depotDonnees.lisParcoursUtilisateur(idUtilisateurCourant);
-      parcoursUtilisateur.etatVisiteGuidee.metEnPause();
-      await depotDonnees.sauvegardeParcoursUtilisateur(parcoursUtilisateur);
-
-      reponse.sendStatus(200);
-    }
-  );
-
-  routes.post(
-    '/visiteGuidee/reprends',
-    middleware.verificationAcceptationCGU,
-    async (requete, reponse) => {
-      const { idUtilisateurCourant } = requete;
-
-      const parcoursUtilisateur =
-        await depotDonnees.lisParcoursUtilisateur(idUtilisateurCourant);
-      parcoursUtilisateur.etatVisiteGuidee.reprends();
-      await depotDonnees.sauvegardeParcoursUtilisateur(parcoursUtilisateur);
-
-      reponse.sendStatus(200);
     }
   );
